Multiple additions, read the description for more
Gave the player the flappy bird sprites and created a flying animation.
Added a ceiling so that the player cannot go higher than a certain height.
Added a ground with the ground sprite.
Added some line of code in the PlayerMovement script to give the player angular momentum (for aesthetic purposes).